WIP:ctests: Add origin-tracking testcase
We're explicitly avoiding calls to "netplan_parser_load_nullable_overrides()"
here, as that is the quirk we want to get rid of.

We should try getting the same results (all settings written to the correct
files), by just calling netplan_state_update_yaml_hierarchy() and tracking the
origin-files implicitly/internally.