WIP: Use the np_state->_private->dirty_fields structure for origin-file tracking
We track the list of filenames a specific (global) setting appeared in (origin-files) in a PATH like string, which separates the different files in reverse order of appearance, separated by colons.

We use "__UNNAMED" as a placeholder when reading from non-files, e.g. an in memory YAML patch, provided by "netplan set".

When writing back the specific setting we can look at the history of origin-files and decide if it should be written back to the latest file, to a (possibly new) fallback file (e.g. when it only appeared in __UNNAMED), or be dropped.